[Refactor] #191 - 루트뷰를 뷰컨과 합치기
