+++ Wable-iOS/Presentation/Profile/Other/View/Cell/OtherProfileEmptyCell.swift
@@ -1,0 +1,8 @@
+//
+//  OtherProfileEmptyCell.swift
+//  Wable-iOS
+//
+//  Created by 김진웅 on 5/20/25.
+//
+
+import Foundation
