+++ Wable-iOS/Presentation/Profile/Other/View/OtherProfileViewController.swift
@@ -1,0 +1,8 @@
+//
+//  OtherProfileViewController.swift
+//  Wable-iOS
+//
+//  Created by 김진웅 on 5/20/25.
+//
+
+import Foundation
