+++ Wable-iOS/Presentation/Profile/Other/ViewModel/OtherProfileViewModel.swift
@@ -1,0 +1,8 @@
+//
+//  OtherProfileViewModel.swift
+//  Wable-iOS
+//
+//  Created by 김진웅 on 5/20/25.
+//
+
+import Foundation
